# Conflicts:
#	src/main/resources/static/asset-manifest.json
#	src/main/resources/static/index.html

+++ src/main/java/com/capstone/project/jwt/repository/RefreshTokenRepository.java
@@ -5,15 +5,15 @@ import com.capstone.project.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Integer> {
 
     // 특정 사용자에 대해 리프레시 토큰을 찾기 위한 메소드
-    Optional<RefreshToken> findByMemberId(Long memberId);
+    Optional<RefreshToken> findByMemberId(Integer memberId);
     Optional<RefreshToken> findByMember(Member member);
     Optional<RefreshToken> findByRefreshToken(String token);
 
     // 리프레시 토큰을 갱신하기 위한 메소드
-    void deleteByMemberId(Long memberId);  // 특정 사용자의 리프레시 토큰 삭제
+    void deleteByMemberId(Integer memberId);  // 특정 사용자의 리프레시 토큰 삭제
 
 }
